+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 
+## [0.6.0] - 2021-01-14
+### Added
+- Install UUID package
+- Feature to archived translations
+- Feature auto-archiving previous translations
+
+### Changed
+- Rename `TranslationNotFoundEvent` into `TranslationNotFound`
+- Rename `TranslationSavedEvent` into `TranslationCreated`
+- Make 'locale' field nullable
+
 ## [0.5.0] - 2021-01-13
 ### Added
 - Laravel 8 support